some more refactoring so that tier 2 SDKs do not have to execute the same logic
