--- src/main/java/com/wavefront/sdk/common/ConnectionHandler.java
@@ -1,40 +1,0 @@
-package com.wavefront.sdk.common;
-
-import java.io.Closeable;
-import java.io.IOException;
-
-/**
- * Connection Handler client that sends data to a Wavefront proxy or
- * Wavefront direct ingestion service.
- *
- * @author Vikram Raman (vikram@wavefront.com).
- */
-public interface ConnectionHandler extends Closeable {
-
-  /**
-   * Connects to the server.
-   *
-   * @throws IllegalStateException if the client is already connected
-   * @throws IOException           if there is an error connecting
-   */
-  void connect() throws IllegalStateException, IOException;
-
-  /**
-   * Flushes buffer, if applicable
-   *
-   * @throws IOException
-   */
-  void flush() throws IOException;
-
-  /**
-   * Returns true if ready to send data
-   */
-  boolean isConnected();
-
-  /**
-   * Returns the number of failed writes to the server.
-   *
-   * @return the number of failed writes to the server
-   */
-  int getFailureCount();
-}
